Move Keyword model from model.commercial to commercial.model.capi
