+++ commercial/app/model/merchandise/books/BestsellersAgent.scala
@@ -1,10 +1,10 @@
 package commercial.model.merchandise.books
 
 import commercial.model.Segment
+import commercial.model.capi.Keyword
 import commercial.model.feeds.{FeedMetaData, ParsedFeed}
 import common.ExecutionContexts
 import commercial.model.merchandise.{Book, MerchandiseAgent}
-import model.commercial.Keyword
 
 import scala.concurrent.Future
