+++ commercial/app/model/merchandise/jobs/JobsAgent.scala
@@ -1,10 +1,10 @@
 package commercial.model.merchandise.jobs
 
 import commercial.model.Segment
+import commercial.model.capi.Keyword
 import commercial.model.feeds.{FeedMetaData, ParsedFeed}
 import common.ExecutionContexts
 import commercial.model.merchandise.{Job, MerchandiseAgent}
-import model.commercial.Keyword
 
 import scala.concurrent.Future
